Front-End (IVR system)/requirements.txt
mysql-connector-python
pyttsx3
SpeechRecognition
pyaudio